feat: 로그아웃 시 앱 아이콘 default 아이콘으로 변경되도록 구현
