--- domain/src/main/java/com/devndev/lamp/domain/manager/AppIconManager.kt
@@ -1,6 +1,0 @@
-package com.devndev.lamp.domain.manager
-
-interface AppIconManager {
-    fun updateAppIcon(gender: String)
-    fun saveIconPreference(gender: String)
-}
