+++ presentation/src/main/java/com/devndev/lamp/presentation/ui/home/HomeViewModel.kt
@@ -4,9 +4,9 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import coil.network.HttpException
-import com.devndev.lamp.domain.manager.AppIconManager
 import com.devndev.lamp.domain.model.MyInfoDomainModel
 import com.devndev.lamp.domain.usecase.GetMyInfoUseCase
+import com.devndev.lamp.presentation.ui.utils.IconStatusManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -15,7 +15,6 @@ import javax.inject.Inject
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
-    private val appIconManager: AppIconManager,
     private val getMyInfoUseCase: GetMyInfoUseCase
 ) : ViewModel() {
     private val logTag = "HomeViewModel"
@@ -34,7 +33,7 @@ class HomeViewModel @Inject constructor(
                 _myInfo.value = getMyInfoUseCase()
                 Log.d(logTag, "My Info ${myInfo.value}")
                 _myInfo.value?.gender?.let {
-                    appIconManager.saveIconPreference(it)
+                    IconStatusManager.setIconStatus(it)
                 }
             } catch (e: HttpException) {
                 Log.e(logTag, "fetchData HttpException", e)
